fixed: output overlaid video has same number of frames as input frames, and will not skip any frames while processing the video if using default sampling frame rate

+++ vision/shared/ProgramOptions.h
@@ -16,7 +16,7 @@ struct ProgramOptionsVideo {
     affdex::Path data_dir;
     affdex::Path input_video_path;
     affdex::Path output_video_path;
-    unsigned int sampling_frame_rate;
+    float sampling_frame_rate;
     bool draw_display;
     unsigned int num_faces;
     bool loop = false;
